SNOW-1599877 Move TestMethod Class to Java Test Dir (#136)
move TestMethod to the correct location
