+++ src/test/java/com/snowflake/snowpark_test/TestMethod.java
@@ -1,6 +1,6 @@
-package com.snowflake.snowpark;
+package com.snowflake.snowpark_test;
 
 @FunctionalInterface
 public interface TestMethod {
-    void run();
+  void run();
 }
